# Fix `MessageEmoteCollection` + additional fixes/improvements