--- a/TwitchLib.Client/Events/OnExistingUsersDetectedArgs.cs
+++ b/TwitchLib.Client/Events/OnExistingUsersDetectedArgs.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 
 namespace TwitchLib.Client.Events
 {
@@ -13,7 +14,7 @@ namespace TwitchLib.Client.Events
         /// <summary>
         /// Property representing string list of existing users.
         /// </summary>
-        public string[] Users;
+        public List<string> Users;
         /// <summary>
         /// Property representing channel bot is connected to.
         /// </summary>