### Fixes
- Fix a bug in HandleNotice where `UnaccountedFor(...)` fallback was called even when `OnUnaccountedFor` event had a handler
- Fix comment on `IrcMessage` ctor
- Fix bug in `MesageEmoteCollection.Add(...)` which was indefinitely pattern + emoteText concatenating replacement regex until running out of maximum string length and crashing the application
- Fix usage of `SortedList` in `MessageEmoteCollection` - replaced with `Dictionary`
- Fix missed slicing in IrcParser to remove `@` at the start of tag sequence
- Fallback to unaccounted for on failed login if no failed login handler is registered (previously it would have been silently ignored)

### Cleanup
- Rewrite `HandleIrcMessageAsync` to use switch expression and forward the task
- Simplify and clean up `HandlePrivMsg`
- Rewrite `HandleNotice` to use switch expression and forward the task

### Performance
#### Environment
```
BenchmarkDotNet=v0.13.5, OS=macOS 14.0 [Darwin 23.0.0]
Apple M1 Pro, 1 CPU, 8 logical and 8 physical cores
.NET SDK=8.0.100-preview.7.23327.3
  [Host]   : .NET 7.0.1 (7.0.122.56804), Arm64 RyuJIT AdvSIMD
  ShortRun : .NET 7.0.1 (7.0.122.56804), Arm64 RyuJIT AdvSIMD

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  
```
#### Before
|               Method |     Mean |    Error |    StdDev |      Gen0 |      Gen1 |      Gen2 | Allocated |
|--------------------- |---------:|---------:|----------:|----------:|----------:|----------:|----------:|
| ClientScenarioCreate | 2.172 ms | 5.363 ms | 0.2940 ms | 1661.1328 | 1656.2500 | 1655.2734 |  24.59 MB |

#### After  (400x faster*)
|               Method |     Mean |     Error |    StdDev |   Gen0 |   Gen1 | Allocated |
|--------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
| ClientScenarioCreate | 5.381 us | 0.2274 us | 0.0125 us | 3.2730 | 0.1144 |  20.06 KB |

*this is illustrative of any moderately long running application which keeps the same instance of `TwitchClient` (and `_channelEmotes`). I have no idea if there have been any complaints but there is no way this was working as intended previously without crashing applications the moment they connect to a big streamer chat stream or if they are just kept running for some time.

Closes #231