@neon-sunset 2023-07-02
@Syzuna I'd like to know your opinion on changing the behavior for failed login - right now it would just fail silently if no handler is registered for `OnIncorrectLogin` event. I changed it to fallback to `OnUnaccountedFor` -> `UnaccountedFor` but this might not be desirable or intended according to how the library is used regularly in user code, which I'm not familiar with.