@Syzuna 2023-07-02
> 

That's a tough one.
originally UnaccountedFor was to catch new stuff Twitch sneaked into IRC that we might want to handle and not to indicate that there was a message an user didnt handle